fix(settings): Kalender-Abos aus dem adminOnly-Blatt in die persoenliche Domaene
Der Server behandelt ICS-Abos seit jeher eigentuemerbasiert: `getAll()` liefert
`shared = 1 OR created_by = ich`, und PATCH/DELETE/sync antworten 403 fuer
fremde Abos - `isAdmin` ist dort ein ZUSATZrecht, keine Voraussetzung. Der
einmalige Kalenderimport traegt ebenfalls kein Admin-Gate. Erreichbar war beides
nur ueber `sync-calendar`, und das Blatt ist adminOnly: in einem
fuenfkoepfigen Haushalt durften vier Mitglieder ein Abo anlegen und kamen nicht
an die Seite - ohne Fehlermeldung, weil findSettingsLeaf fuer sie null liefert.

Beides zieht auf ein eigenes Blatt unter Persoenlich. Auf `sync-calendar`
bleibt, was wirklich dem Haushalt gehoert: CalDAV-Konten und die
Google-/Apple-Anbindung, deren Routen `requireAdmin` tragen.

Fuenfter Fall desselben Musters nach den Kalender-Vorgaben, den
Aufgaben-Vorgaben (#695), der Navigationsreihenfolge und den Feed-Abos - und
der zweite, den nicht das Lesen von Seiten gefunden hat, sondern das Messen der
Regel gegen den Server.

`loadFamilyUsers` wird jetzt von beiden Blaettern gebraucht und liegt deshalb in
`public/settings/family-users.js`: zwei Kopien haetten zwei Zwischenspeicher
bedeutet und damit zwei Abfragen fuer dieselbe Antwort.

Dazu ein Fund aus dem Precache-Guard: SECHS von 28 Settings-Blaettern waren gar
nicht precacht, vier davon seit laengerem (admin-email, admin-permissions,
personal-health, personal-weather). Offline liessen sie sich nicht oeffnen, und
zwar stumm. Der Modulgraph-Guard kann das prinzipiell nicht sehen - er folgt
Importen, ein Blatt haengt aber an einem dynamischen `import()` und steht in
keinem statischen Importbaum. Ein zweiter Guard haelt jetzt die Regel, die er
nicht sieht, und misst gegen die Registry statt gegen das Verzeichnis, damit
ein totes Blatt nicht als fehlend zaehlt. Gegengeprueft: zwei Eintraege
entfernt, er nennt beide beim Namen.
