+++ public/locales/ar.json
@@ -2017,6 +2017,8 @@
         "pageNavigationDescription": "حدّد ترتيب الوحدات والتنقل على الهاتف.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "المطبخ",
         "pageKitchenDescription": "اختر الوجبات الظاهرة في خطة الوجبات.",
         "pageModuleOptions": "خيارات الوحدات",
