+++ public/locales/cs.json
@@ -2017,6 +2017,8 @@
         "pageNavigationDescription": "Nastavte pořadí modulů a mobilní navigaci.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "Kuchyně",
         "pageKitchenDescription": "Vyberte, která jídla se zobrazí v jídelníčku.",
         "pageModuleOptions": "Volby modulů",
