+++ public/locales/hi.json
@@ -2017,6 +2017,8 @@
         "pageNavigationDescription": "मॉड्यूल का क्रम और मोबाइल नेविगेशन सेट करें।",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "रसोई",
         "pageKitchenDescription": "चुनें कि भोजन योजना में कौन-से भोजन दिखें।",
         "pageModuleOptions": "मॉड्यूल विकल्प",
