+++ public/locales/id.json
@@ -2029,6 +2029,8 @@
         "pageNavigationDescription": "Atur urutan modul dan navigasi seluler.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "Dapur",
         "pageKitchenDescription": "Pilih makanan mana yang tampil di rencana makan.",
         "pageModuleOptions": "Opsi modul",
