+++ public/locales/it.json
@@ -2017,6 +2017,8 @@
         "pageNavigationDescription": "Imposta l'ordine dei moduli e la navigazione mobile.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "Cucina",
         "pageKitchenDescription": "Scegli quali pasti sono visibili nel piano pasti.",
         "pageModuleOptions": "Opzioni dei moduli",
