+++ public/locales/ko.json
@@ -2029,6 +2029,8 @@
         "pageNavigationDescription": "모듈 순서와 모바일 내비게이션을 설정합니다.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "주방",
         "pageKitchenDescription": "식단 계획에 표시할 식사를 선택합니다.",
         "pageModuleOptions": "모듈 옵션",
