+++ public/locales/pt.json
@@ -2017,6 +2017,8 @@
         "pageNavigationDescription": "Defina a ordem dos módulos e a navegação móvel.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "Cozinha",
         "pageKitchenDescription": "Escolha quais refeições ficam visíveis no plano de refeições.",
         "pageModuleOptions": "Opções de módulos",
