+++ public/locales/sv.json
@@ -2017,6 +2017,8 @@
         "pageNavigationDescription": "Ange modulernas ordning och den mobila navigeringen.",
         "pageFeeds": "Feed subscriptions",
         "pageFeedsDescription": "Subscribe to the household calendar and inventory deadlines in your own calendar app.",
+        "pageCalendarSubscriptions": "Calendar subscriptions",
+        "pageCalendarSubscriptionsDescription": "Subscribe to ICS/webcal calendars and import events once from a file or URL.",
         "pageKitchen": "Kök",
         "pageKitchenDescription": "Välj vilka måltider som visas i måltidsplanen.",
         "pageModuleOptions": "Modulalternativ",
