+++ public/sw.js
@@ -227,7 +227,14 @@ const PAGE_MODULES = [
   '/settings/preferences-cache.js',
   '/settings/region-presets.js',
   '/settings/weather-location.js',
+  '/settings/family-users.js',
   '/settings/pages/personal-account.js',
+  '/settings/pages/admin-email.js',
+  '/settings/pages/admin-permissions.js',
+  '/settings/pages/personal-calendar-subscriptions.js',
+  '/settings/pages/personal-feeds.js',
+  '/settings/pages/personal-health.js',
+  '/settings/pages/personal-weather.js',
   '/settings/pages/personal-appearance.js',
   '/settings/pages/personal-device.js',
   '/settings/pages/personal-calendar.js',
