+++ test/test-settings-navigation.js
@@ -118,11 +118,16 @@ test('die Blätter verteilen sich wie beschlossen auf die vier Domänen', () => 
   // sehen will, entscheide ich.
   // Und ebenso `personal-feeds`: beide Feed-Tokens hängen an der eigenen
   // users-Zeile und beide Routen tragen keinen Admin-Check, das Blatt lag
-  // trotzdem im adminOnly-`sync-calendar`. Was in den Kalender hineinkommt
-  // (CalDAV, ICS-Abos, Import) bleibt dort.
+  // trotzdem im adminOnly-`sync-calendar`.
+  // `personal-calendar-subscriptions` ist die Gegenrichtung und derselbe Fall:
+  // `GET /calendar/subscriptions` liefert `shared = 1 OR created_by = ich`, und
+  // PATCH/DELETE/sync antworten 403 für fremde Abos - `isAdmin` ist dort ein
+  // ZUSATZrecht, keine Voraussetzung. Bei `sync` bleiben nur die Blätter, deren
+  // Routen wirklich `requireAdmin` tragen: CalDAV und Google/Apple hängen an
+  // Zugangsdaten des Haushalts.
   const perDomain = {};
   for (const leaf of SETTINGS_LEAVES) perDomain[leaf.domainId] = (perDomain[leaf.domainId] ?? 0) + 1;
-  assert.deepEqual(perDomain, { personal: 10, modules: 4, sync: 5, admin: 8 });
+  assert.deepEqual(perDomain, { personal: 11, modules: 4, sync: 5, admin: 8 });
   // Jedes Blatt hängt an einer existierenden Domäne.
   const domainIds = new Set(SETTINGS_DOMAINS.map((domain) => domain.id));
   for (const leaf of SETTINGS_LEAVES) {
