+++ test/test-sw-precache.js
@@ -166,3 +166,30 @@ test('keine Doppeleinträge zwischen den Precache-Listen', () => {
   const dupes = all.filter((p, i) => all.indexOf(p) !== i);
   assert.deepEqual([...new Set(dupes)], [], `Mehrfach precacht: ${dupes.join(', ')}`);
 });
+
+test('jedes Settings-Blatt der Registry ist precacht', () => {
+  // Schwesterregel zum Modulgraph-Guard oben, und die Lücke, die er offen
+  // lässt: er folgt Importen ab den precachten Modulen, ein Blatt aber wird
+  // per dynamischem `loader: () => import(...)` geladen und steht damit in
+  // keinem statischen Importbaum. Ein nicht precachtes Blatt ist deshalb kein
+  // Mischzustand, sondern schlicht offline nicht erreichbar - stumm, weil es
+  // online immer geht. Gemessen am 2026-08-15 fehlten sechs von 28, vier davon
+  // seit längerem: admin-email, admin-permissions, personal-health und
+  // personal-weather.
+  //
+  // Kanonische Quelle ist die Registry, nicht das Verzeichnis: ein Blatt, das
+  // dort nicht steht, ist tot und muss nicht precacht sein.
+  const registry = readFileSync(new URL('../public/settings/registry.js', import.meta.url), 'utf8');
+  const leaves = [...registry.matchAll(/loader:\s*\(\)\s*=>\s*import\('([^']+)'\)/g)].map((m) => m[1]);
+
+  assert.ok(leaves.length >= 20,
+    `Nur ${leaves.length} Blätter in der Registry gefunden - das Muster greift nicht mehr`);
+
+  const precached = new Set([...APP_SHELL, ...PAGE_MODULES]);
+  const missing = leaves.filter((path) => !precached.has(path));
+  assert.deepEqual(
+    missing, [],
+    'Diese Settings-Blätter stehen in der Registry, werden aber nicht precacht '
+    + `und sind damit offline nicht erreichbar:\n  ${missing.join('\n  ')}`,
+  );
+});
